libInternal/db.py
import os
from sqlalchemy import create_engine
from urllib.parse import quote_plus
from dotenv import load_dotenv

def getConnection():
    load_dotenv()
    db_host = os.getenv("DB_HOST")
    db_user = os.getenv("DB_USER")
    db_pass = quote_plus(os.getenv("DB_PASS"))
    db_name = os.getenv("DB_NAME")
    return create_engine(f"mysql+mysqlconnector://{db_user}:{db_pass}@{db_host}/{db_name}")