requirements.txt
pandas
numpy
duckdb
sqlalchemy
python-dotenv
plotly
scikit-learn
networkx
flask
tqdm
